feat(dtt-29): no content-length, url not fileID
Changed parcel to use URLs instead of fileIDs. This means parcel is now
more of a general tool rather than an extension of the gdc-client.

Added a segment to take care of downloads that don't have a
Content-Length in the header. This means that parallel chunked downloads
are not possible, but we can still download the files.

fix(review): fixed comment that didn't reflect code

fix(review): changed error message to give more information